Move breadcrumbs link generation logic to breadcrumbs service.
